backend/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
passlib[bcrypt]
pydantic
pyjwt
supabase